examples/py-sdk-demo/pyproject.toml
[project]
name = "py-sdk-demo"
version = "0.1.0"
description = "HX-HouTiKu Python SDK 使用示例"
requires-python = ">=3.10"
dependencies = [
    "hx-houtiku",
    "python-dotenv>=1.0",
]
